moved queue related trait to own folder
